requirements.txt
# E-GAN Radio Anomaly Detection - Dependencies
# Python 3.8+ recommended

# Core
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
scipy>=1.7.0

# Configuration
pyyaml>=6.0
omegaconf>=2.1.0

# Data processing
pandas>=1.3.0
tqdm>=4.62.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Metrics
scikit-learn>=1.0.0

# Audio/Signal processing (optional, for wav fallback)
soundfile>=0.10.0

# Image processing for spectrogram resize
pillow>=8.0.0
